+++ main-service/src/main/java/ru/practicum/main/location/repository/LocationRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import ru.practicum.main.location.model.Location;
 
 public interface LocationRepository extends JpaRepository<Location, Long> {
-    Optional<Location> findByLonAndLat(Double lon, Double lat);
+    Optional<Location> findByLonAndLat(Float lon, Float lat);
 }
